Bug 1141850 - Part 1: rename MirroringBookmarksSynchronizer to BufferingBookmarksSynchronizer.
